# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PerceptChain Backend API",
    "description": "API for fetching and visualizing Solana blockchain data using Helius RPC",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://perceptchain-backend-web.vercel.app/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/token-holders/{token_address}/{limit}": {
      "get": {
        "summary": "Get top token holders",
        "description": "Retrieves the top holders of a specific token by amount",
        "operationId": "getTokenHolders",
        "parameters": [
          {
            "name": "token_address",
            "in": "path",
            "description": "The mint address of the token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
          },
          {
            "name": "limit",
            "in": "path",
            "description": "Maximum number of holders to return",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "address": {
                        "type": "string",
                        "description": "The wallet address of the holder"
                      },
                      "amount": {
                        "type": "string",
                        "description": "Raw token amount as a string"
                      },
                      "uiAmount": {
                        "type": "number",
                        "description": "Human-readable token amount accounting for decimals"
                      },
                      "percentage": {
                        "type": "number",
                        "format": "float",
                        "description": "Percentage of total supply held"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Invalid token mint address: xyz"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Gateway timeout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Helius API timed out"
                }
              }
            }
          }
        }
      }
    },
    "/assets/{group_key}/{group_value}": {
      "get": {
        "summary": "Get assets by group",
        "description": "Retrieves assets by collection, creator, or owner with analytics",
        "operationId": "getAssetsByGroup",
        "parameters": [
          {
            "name": "group_key",
            "in": "path",
            "description": "The type of grouping (collection, creator, owner)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["collection", "creator", "owner"]
            },
            "example": "collection"
          },
          {
            "name": "group_value",
            "in": "path",
            "description": "The value to search for (collection ID, creator address, owner address)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "4mKSoDDqApmF1DqXvVTSL6tu2zixrSSNjqMxUnwvVzy2"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "Total number of assets in the group"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Number of assets per page"
                    },
                    "page": {
                      "type": "integer",
                      "description": "Current page number"
                    },
                    "groupKey": {
                      "type": "string",
                      "description": "Type of grouping used"
                    },
                    "groupValue": {
                      "type": "string",
                      "description": "Value used for grouping"
                    },
                    "assets": {
                      "type": "array",
                      "description": "List of assets",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Asset ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "Asset name"
                          },
                          "symbol": {
                            "type": "string",
                            "description": "Asset symbol"
                          },
                          "image": {
                            "type": "string",
                            "description": "Asset image URL"
                          },
                          "owner": {
                            "type": "string",
                            "description": "Current owner address"
                          },
                          "attributes": {
                            "type": "array",
                            "description": "Asset attributes/traits",
                            "items": {
                              "type": "object"
                            }
                          },
                          "royalty": {
                            "type": "number",
                            "description": "Royalty percentage"
                          },
                          "collection": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "description": "Collection name"
                              },
                              "id": {
                                "type": "string",
                                "description": "Collection ID"
                              }
                            }
                          }
                        }
                      }
                    },
                    "attribute_stats": {
                      "type": "object",
                      "description": "Statistics about attribute distribution"
                    },
                    "owner_distribution": {
                      "type": "object",
                      "description": "Statistics about ownership distribution"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/wallet/tokens/{wallet_address}": {
      "get": {
        "summary": "Get token accounts by owner",
        "description": "Retrieves all SPL token accounts owned by a specific wallet",
        "operationId": "getTokenAccountsByOwner",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "description": "The wallet address to query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "DezXAZ8z7PnrnRJjz3wXBoRgixCa6xjnB7YaB1pPB263"
          },
          {
            "name": "include_details",
            "in": "query",
            "description": "Whether to include detailed token metadata",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "owner": {
                      "type": "string",
                      "description": "Wallet address"
                    },
                    "count": {
                      "type": "integer",
                      "description": "Number of token accounts"
                    },
                    "tokens": {
                      "type": "array",
                      "description": "List of token accounts",
                      "items": {
                        "type": "object",
                        "properties": {
                          "mint": {
                            "type": "string",
                            "description": "Token mint address"
                          },
                          "address": {
                            "type": "string",
                            "description": "Token account address"
                          },
                          "amount": {
                            "type": "string",
                            "description": "Raw token amount as a string"
                          },
                          "uiAmount": {
                            "type": "number",
                            "description": "Human-readable token amount"
                          },
                          "decimals": {
                            "type": "integer",
                            "description": "Number of decimal places"
                          },
                          "tokenSupply": {
                            "type": "object",
                            "description": "Token supply information",
                            "properties": {
                              "amount": {
                                "type": "string",
                                "description": "Raw total supply"
                              },
                              "uiAmount": {
                                "type": "number",
                                "description": "Human-readable total supply"
                              },
                              "decimals": {
                                "type": "integer",
                                "description": "Number of decimal places"
                              }
                            }
                          },
                          "percentageOwned": {
                            "type": "number",
                            "description": "Percentage of total supply owned"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid wallet address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{address}": {
      "get": {
        "summary": "Get signatures for address",
        "description": "Retrieves transaction signatures for an address with analytics",
        "operationId": "getSignaturesForAddress",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The wallet or contract address to query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "DezXAZ8z7PnrnRJjz3wXBoRgixCa6xjnB7YaB1pPB263"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of signatures to fetch",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Start searching from this signature (for pagination)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "Search until this signature",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string",
                      "description": "Address queried"
                    },
                    "count": {
                      "type": "integer",
                      "description": "Number of signatures returned"
                    },
                    "successRate": {
                      "type": "number",
                      "description": "Percentage of successful transactions"
                    },
                    "totalFees": {
                      "type": "integer",
                      "description": "Total transaction fees in lamports"
                    },
                    "signatures": {
                      "type": "array",
                      "description": "List of transaction signatures",
                      "items": {
                        "type": "object",
                        "properties": {
                          "signature": {
                            "type": "string",
                            "description": "Transaction signature"
                          },
                          "blockTime": {
                            "type": "integer",
                            "description": "Block timestamp (Unix epoch)"
                          },
                          "readableTime": {
                            "type": "string",
                            "description": "Human-readable timestamp"
                          },
                          "slot": {
                            "type": "integer",
                            "description": "Slot number"
                          },
                          "err": {
                            "type": "object",
                            "nullable": true,
                            "description": "Error information if transaction failed"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["success", "failed"],
                            "description": "Transaction status"
                          },
                          "fee": {
                            "type": "integer",
                            "description": "Transaction fee in lamports"
                          },
                          "memo": {
                            "type": "string",
                            "nullable": true,
                            "description": "Transaction memo"
                          }
                        }
                      }
                    },
                    "analytics": {
                      "type": "object",
                      "properties": {
                        "hourlyActivity": {
                          "type": "array",
                          "description": "Transaction count by hour of day",
                          "items": {
                            "type": "object",
                            "properties": {
                              "hour": {
                                "type": "integer",
                                "description": "Hour (0-23)"
                              },
                              "count": {
                                "type": "integer",
                                "description": "Number of transactions"
                              }
                            }
                          }
                        },
                        "topPrograms": {
                          "type": "array",
                          "description": "Most frequently used programs",
                          "items": {
                            "type": "object",
                            "properties": {
                              "program": {
                                "type": "string",
                                "description": "Program address"
                              },
                              "count": {
                                "type": "integer",
                                "description": "Number of invocations"
                              }
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "before": {
                          "type": "string",
                          "nullable": true,
                          "description": "Signature to use for the next page"
                        },
                        "hasMore": {
                          "type": "boolean",
                          "description": "Whether more results are available"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Gateway timeout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}